@@ -11,33 +11,34 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "next": "14.2.30",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "wagmi": "^2.5.0",
-    "viem": "^2.7.0",
-    "@rainbow-me/rainbowkit": "^2.0.0",
-    "ethers": "^6.8.0",
-    "zustand": "^4.4.0",
-    "react-hook-form": "^7.47.0",
-    "zod": "^3.22.0",
     "@hookform/resolvers": "^3.3.0",
-    "clsx": "^2.0.0",
+    "@partisipro/shared": "file:../../packages/shared",
+    "@rainbow-me/rainbowkit": "^2.0.0",
     "class-variance-authority": "^0.7.0",
-    "lucide-react": "^0.292.0",
-    "recharts": "^2.8.0",
+    "clsx": "^2.0.0",
     "date-fns": "^2.30.0",
-    "@partisipro/shared": "file:../../packages/shared"
+    "ethers": "^6.8.0",
+    "lucide-react": "^0.292.0",
+    "next": "^15.1.0",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-hook-form": "^7.47.0",
+    "recharts": "^2.8.0",
+    "viem": "^2.7.0",
+    "wagmi": "^2.5.0",
+    "zod": "^3.22.0",
+    "zustand": "^4.4.0"
   },
   "devDependencies": {
-    "@types/node": "20.8.0",
-    "@types/react": "18.2.0",
-    "@types/react-dom": "18.2.0",
-    "autoprefixer": "10.4.0",
-    "eslint": "8.50.0",
-    "eslint-config-next": "14.0.0",
-    "postcss": "8.5.6",
-    "tailwindcss": "3.3.0",
-    "typescript": "5.2.2"
+    "@types/node": "22.10.2",
+    "@types/react": "19.0.2",
+    "@types/react-dom": "19.0.2",
+    "autoprefixer": "10.4.20",
+    "eslint": "8.57.1",
+    "eslint-config-next": "15.1.0",
+    "postcss": "8.4.49",
+    "tailwindcss": "^4.0.0",
+    "@tailwindcss/postcss": "^4.0.0",
+    "typescript": "5.7.2"
   }
 }
